I think i cleaned up the most of the native methods
