--- External Functions/Program.cs
@@ -1,141 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-using System.Windows.Interop;
-using System.Windows.Media.Imaging;
-using SchoolWallpaperChanger.External_Functions;
-
-namespace DrawBehindDesktopIcons
-{
-    class Program
-    {
-
-        [DllImport("User32.dll")]
-        static extern IntPtr GetDC(IntPtr hwnd);
-
-        [DllImport("User32.dll")]
-        static extern int ReleaseDC(IntPtr hwnd, IntPtr dc);
-        public static IntPtr workerw = IntPtr.Zero;
-        public static void Start(string pictureLocation)
-        {
-            PrintVisibleWindowHandles(2);
-            // The output will look something like this. 
-            // .....
-            // 0x00010190 "" WorkerW
-            //   ...
-            //   0x000100EE "" SHELLDLL_DefView
-            //     0x000100F0 "FolderView" SysListView32
-            // 0x000100EC "Program Manager" Progman
-
-
-
-            // Fetch the Progman window
-            IntPtr progman = W32.FindWindow("Progman", null);
-
-            IntPtr result = IntPtr.Zero;
-
-            // Send 0x052C to Progman. This message directs Progman to spawn a 
-            // WorkerW behind the desktop icons. If it is already there, nothing 
-            // happens.
-            W32.SendMessageTimeout(progman,
-                                   0x052C,
-                                   new IntPtr(0),
-                                   IntPtr.Zero,
-                                   W32.SendMessageTimeoutFlags.SMTO_NORMAL,
-                                   1000,
-                                   out result);
-
-
-            PrintVisibleWindowHandles(2);
-            // The output will look something like this
-            // .....
-            // 0x00010190 "" WorkerW
-            //   ...
-            //   0x000100EE "" SHELLDLL_DefView
-            //     0x000100F0 "FolderView" SysListView32
-            // 0x00100B8A "" WorkerW                                   <--- This is the WorkerW instance we are after!
-            // 0x000100EC "Program Manager" Progman
-
-            
-
-            // We enumerate all Windows, until we find one, that has the SHELLDLL_DefView 
-            // as a child. 
-            // If we found that window, we take its next sibling and assign it to workerw.
-            W32.EnumWindows(new W32.EnumWindowsProc((tophandle, topparamhandle) =>
-            {
-                IntPtr p = W32.FindWindowEx(tophandle,
-                                            IntPtr.Zero,
-                                            "SHELLDLL_DefView",
-                                            IntPtr.Zero);
-
-                if (p != IntPtr.Zero)
-                {
-                    // Gets the WorkerW Window after the current one.
-                    workerw = W32.FindWindowEx(IntPtr.Zero,
-                                               tophandle,
-                                               "WorkerW",
-                                               IntPtr.Zero);
-                }
-
-                return true;
-            }), IntPtr.Zero);
-            ScreenRes.SetDpiAwareness();
-            Window1 window1 = new Window1();
-            window1.Show();
-            BitmapImage bitmap = new BitmapImage();
-            bitmap.BeginInit();
-            bitmap.UriSource = new Uri(pictureLocation);
-            bitmap.CacheOption = BitmapCacheOption.OnLoad;
-            bitmap.EndInit();
-            window1.image.Source = bitmap;
-        }
-        public static void RemoveWindowFromTaskbar(IntPtr handle)
-        {
-            var styleNewWindowExtended =
-                   (Int64)NativeMethods.WindowStyles.WS_EX_NOACTIVATE
-                   | (Int64)NativeMethods.WindowStyles.WS_EX_TOOLWINDOW;
-
-            // update window styles
-            NativeMethods.SetWindowLongPtr(new HandleRef(null, handle), (-20), (IntPtr)styleNewWindowExtended);
-        }
-        static void PrintVisibleWindowHandles(IntPtr hwnd, int maxLevel = -1, int level = 0)
-        {
-            bool isVisible = W32.IsWindowVisible(hwnd);
-
-            if (isVisible && (maxLevel == -1 || level <= maxLevel))
-            {
-                StringBuilder className = new StringBuilder(256);
-                W32.GetClassName(hwnd, className, className.Capacity);
-
-                StringBuilder windowTitle = new StringBuilder(256);
-                W32.GetWindowText(hwnd, windowTitle, className.Capacity);
-
-                Console.WriteLine("".PadLeft(level * 2) + "0x{0:X8} \"{1}\" {2}", hwnd.ToInt64(), windowTitle, className);
-
-                level++;
-
-                // Enumerates all child windows of the current window
-                W32.EnumChildWindows(hwnd, new W32.EnumWindowsProc((childhandle, childparamhandle) =>
-                {
-                    PrintVisibleWindowHandles(childhandle, maxLevel, level);
-                    return true;
-                }), IntPtr.Zero);
-            }
-        }
-        static void PrintVisibleWindowHandles(int maxLevel = -1)
-        {
-            // Enumerates all existing top window handles. This includes open and visible windows, as well as invisible windows.
-            W32.EnumWindows(new W32.EnumWindowsProc((tophandle, topparamhandle) =>
-            {
-                PrintVisibleWindowHandles(tophandle, maxLevel);
-                return true;
-            }), IntPtr.Zero);
-        }
-    }
-}
