--- External Functions/Win32Wrapper.cs
@@ -1,729 +1,0 @@
-﻿using Microsoft.Win32;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-
-namespace DrawBehindDesktopIcons
-{
-    public class W32
-    {
-        public const int SM_CXSCREEN = 0;
-
-        public const int SM_CYSCREEN = 1;
-
-        public const int SPI_SETDESKWALLPAPER = 20;
-        public const int SPIF_SENDWININICHANGE = 0x02;
-        public const int SPIF_UPDATEINIFILE = 0x01;
-        public const int SRCCOPY = 13369376;
-
-        public const int WM_GETICON = 0x7F;
-
-        public delegate bool EnumWindowsProc(IntPtr hwnd, IntPtr lParam);
-
-        [Flags]
-        public enum BbFlags : byte //Blur Behind Flags
-        {
-            DwmBbEnable = 1,
-            DwmBbBlurregion = 2,
-            DwmBbTransitiononmaximized = 4,
-        };
-
-        /// <summary>Values to pass to the GetDCEx method.</summary>
-        [Flags()]
-        public enum DeviceContextValues : uint
-        {
-            /// <summary>DCX_WINDOW: Returns a DC that corresponds to the window rectangle rather
-            /// than the client rectangle.</summary>
-            Window = 0x00000001,
-
-            /// <summary>DCX_CACHE: Returns a DC from the cache, rather than the OWNDC or CLASSDC
-            /// window. Essentially overrides CS_OWNDC and CS_CLASSDC.</summary>
-            Cache = 0x00000002,
-
-            /// <summary>DCX_NORESETATTRS: Does not reset the attributes of this DC to the
-            /// default attributes when this DC is released.</summary>
-            NoResetAttrs = 0x00000004,
-
-            /// <summary>DCX_CLIPCHILDREN: Excludes the visible regions of all child windows
-            /// below the window identified by hWnd.</summary>
-            ClipChildren = 0x00000008,
-
-            /// <summary>DCX_CLIPSIBLINGS: Excludes the visible regions of all sibling windows
-            /// above the window identified by hWnd.</summary>
-            ClipSiblings = 0x00000010,
-
-            /// <summary>DCX_PARENTCLIP: Uses the visible region of the parent window. The
-            /// parent's WS_CLIPCHILDREN and CS_PARENTDC style bits are ignored. The origin is
-            /// set to the upper-left corner of the window identified by hWnd.</summary>
-            ParentClip = 0x00000020,
-
-            /// <summary>DCX_EXCLUDERGN: The clipping region identified by hrgnClip is excluded
-            /// from the visible region of the returned DC.</summary>
-            ExcludeRgn = 0x00000040,
-
-            /// <summary>DCX_INTERSECTRGN: The clipping region identified by hrgnClip is
-            /// intersected with the visible region of the returned DC.</summary>
-            IntersectRgn = 0x00000080,
-
-            /// <summary>DCX_EXCLUDEUPDATE: Unknown...Undocumented</summary>
-            ExcludeUpdate = 0x00000100,
-
-            /// <summary>DCX_INTERSECTUPDATE: Unknown...Undocumented</summary>
-            IntersectUpdate = 0x00000200,
-
-            /// <summary>DCX_LOCKWINDOWUPDATE: Allows drawing even if there is a LockWindowUpdate
-            /// call in effect that would otherwise exclude this window. Used for drawing during
-            /// tracking.</summary>
-            LockWindowUpdate = 0x00000400,
-
-            /// <summary>DCX_USESTYLE: Undocumented, something related to WM_NCPAINT message.</summary>
-            UseStyle = 0x00010000,
-
-            /// <summary>DCX_VALIDATE When specified with DCX_INTERSECTUPDATE, causes the DC to
-            /// be completely validated. Using this function with both DCX_INTERSECTUPDATE and
-            /// DCX_VALIDATE is identical to using the BeginPaint function.</summary>
-            Validate = 0x00200000,
-        }
-
-        [Flags]
-        public enum RedrawWindowFlags : uint
-        {
-            /// <summary>
-            /// Invalidates the rectangle or region that you specify in lprcUpdate or hrgnUpdate.
-            /// You can set only one of these parameters to a non-NULL value. If both are NULL, RDW_INVALIDATE invalidates the entire window.
-            /// </summary>
-            Invalidate = 0x1,
-
-            /// <summary>Causes the OS to post a WM_PAINT message to the window regardless of whether a portion of the window is invalid.</summary>
-            InternalPaint = 0x2,
-
-            /// <summary>
-            /// Causes the window to receive a WM_ERASEBKGND message when the window is repainted.
-            /// Specify this value in combination with the RDW_INVALIDATE value; otherwise, RDW_ERASE has no effect.
-            /// </summary>
-            Erase = 0x4,
-
-            /// <summary>
-            /// Validates the rectangle or region that you specify in lprcUpdate or hrgnUpdate.
-            /// You can set only one of these parameters to a non-NULL value. If both are NULL, RDW_VALIDATE validates the entire window.
-            /// This value does not affect internal WM_PAINT messages.
-            /// </summary>
-            Validate = 0x8,
-
-            NoInternalPaint = 0x10,
-
-            /// <summary>Suppresses any pending WM_ERASEBKGND messages.</summary>
-            NoErase = 0x20,
-
-            /// <summary>Excludes child windows, if any, from the repainting operation.</summary>
-            NoChildren = 0x40,
-
-            /// <summary>Includes child windows, if any, in the repainting operation.</summary>
-            AllChildren = 0x80,
-
-            /// <summary>Causes the affected windows, which you specify by setting the RDW_ALLCHILDREN and RDW_NOCHILDREN values, to receive WM_ERASEBKGND and WM_PAINT messages before the RedrawWindow returns, if necessary.</summary>
-            UpdateNow = 0x100,
-
-            /// <summary>
-            /// Causes the affected windows, which you specify by setting the RDW_ALLCHILDREN and RDW_NOCHILDREN values, to receive WM_ERASEBKGND messages before RedrawWindow returns, if necessary.
-            /// The affected windows receive WM_PAINT messages at the ordinary time.
-            /// </summary>
-            EraseNow = 0x200,
-
-            Frame = 0x400,
-
-            NoFrame = 0x800
-        }
-
-        [Flags()]
-        public enum SetWindowPosFlags : uint
-        {
-            /// <summary>If the calling thread and the thread that owns the window are attached to different input queues,
-            /// the system posts the request to the thread that owns the window. This prevents the calling thread from
-            /// blocking its execution while other threads process the request.</summary>
-            /// <remarks>SWP_ASYNCWINDOWPOS</remarks>
-            AsynWindowPos = 0x4000,
-
-            /// <summary>Prevents generation of the WM_SYNCPAINT message.</summary>
-            /// <remarks>SWP_DEFERERASE</remarks>
-            DeferErase = 0x2000,
-
-            /// <summary>Draws a frame (defined in the window's class description) around the window.</summary>
-            /// <remarks>SWP_DRAWFRAME</remarks>
-            DrawFrame = 0x0020,
-
-            /// <summary>Applies new frame styles set using the SetWindowLong function. Sends a WM_NCCALCSIZE message to
-            /// the window, even if the window's size is not being changed. If this flag is not specified, WM_NCCALCSIZE
-            /// is sent only when the window's size is being changed.</summary>
-            /// <remarks>SWP_FRAMECHANGED</remarks>
-            FrameChanged = 0x0020,
-
-            /// <summary>Hides the window.</summary>
-            /// <remarks>SWP_HIDEWINDOW</remarks>
-            HideWindow = 0x0080,
-
-            /// <summary>Does not activate the window. If this flag is not set, the window is activated and moved to the
-            /// top of either the topmost or non-topmost group (depending on the setting of the hWndInsertAfter
-            /// parameter).</summary>
-            /// <remarks>SWP_NOACTIVATE</remarks>
-            NoActivate = 0x0010,
-
-            /// <summary>Discards the entire contents of the client area. If this flag is not specified, the valid
-            /// contents of the client area are saved and copied back into the client area after the window is sized or
-            /// repositioned.</summary>
-            /// <remarks>SWP_NOCOPYBITS</remarks>
-            NoCopyBits = 0x0100,
-
-            /// <summary>Retains the current position (ignores X and Y parameters).</summary>
-            /// <remarks>SWP_NOMOVE</remarks>
-            NoMove = 0x0002,
-
-            /// <summary>Does not change the owner window's position in the Z order.</summary>
-            /// <remarks>SWP_NOOWNERZORDER</remarks>
-            NoOwnerZOrder = 0x0200,
-
-            /// <summary>Does not redraw changes. If this flag is set, no repainting of any kind occurs. This applies to
-            /// the client area, the nonclient area (including the title bar and scroll bars), and any part of the parent
-            /// window uncovered as a result of the window being moved. When this flag is set, the application must
-            /// explicitly invalidate or redraw any parts of the window and parent window that need redrawing.</summary>
-            /// <remarks>SWP_NOREDRAW</remarks>
-            NoRedraw = 0x0008,
-
-            /// <summary>Same as the SWP_NOOWNERZORDER flag.</summary>
-            /// <remarks>SWP_NOREPOSITION</remarks>
-            NoReposition = 0x0200,
-
-            /// <summary>Prevents the window from receiving the WM_WINDOWPOSCHANGING message.</summary>
-            /// <remarks>SWP_NOSENDCHANGING</remarks>
-            NoSendChanging = 0x0400,
-
-            /// <summary>Retains the current size (ignores the cx and cy parameters).</summary>
-            /// <remarks>SWP_NOSIZE</remarks>
-            NoSize = 0x0001,
-
-            /// <summary>Retains the current Z order (ignores the hWndInsertAfter parameter).</summary>
-            /// <remarks>SWP_NOZORDER</remarks>
-            NoZOrder = 0x0004,
-
-            /// <summary>Displays the window.</summary>
-            /// <remarks>SWP_SHOWWINDOW</remarks>
-            ShowWindow = 0x0040
-        }
-
-        public enum WindowLongFlags : int
-        {
-            GWL_EXSTYLE = -20,
-            GWLP_HINSTANCE = -6,
-            GWLP_HWNDPARENT = -8,
-            GWL_ID = -12,
-            GWL_STYLE = -16,
-            GWL_USERDATA = -21,
-            GWL_WNDPROC = -4,
-            DWLP_USER = 0x8,
-            DWLP_MSGRESULT = 0x0,
-            DWLP_DLGPROC = 0x4
-        }
-
-        /// <summary>
-        /// Window Styles.
-        /// The following styles can be specified wherever a window style is required. After the control has been created, these styles cannot be modified, except as noted.
-        /// </summary>
-        [Flags]
-        public enum WindowStyles : uint
-        {
-            /// <summary>The window has a thin-line border.</summary>
-            WS_BORDER = 0x800000,
-
-            /// <summary>The window has a title bar (includes the WS_BORDER style).</summary>
-            WS_CAPTION = 0xc00000,
-
-            /// <summary>The window is a child window. A window with this style cannot have a menu bar. This style cannot be used with the WS_POPUP style.</summary>
-            WS_CHILD = 0x40000000,
-
-            /// <summary>Excludes the area occupied by child windows when drawing occurs within the parent window. This style is used when creating the parent window.</summary>
-            WS_CLIPCHILDREN = 0x2000000,
-
-            /// <summary>
-            /// Clips child windows relative to each other; that is, when a particular child window receives a WM_PAINT message, the WS_CLIPSIBLINGS style clips all other overlapping child windows out of the region of the child window to be updated.
-            /// If WS_CLIPSIBLINGS is not specified and child windows overlap, it is possible, when drawing within the client area of a child window, to draw within the client area of a neighboring child window.
-            /// </summary>
-            WS_CLIPSIBLINGS = 0x4000000,
-
-            /// <summary>The window is initially disabled. A disabled window cannot receive input from the user. To change this after a window has been created, use the EnableWindow function.</summary>
-            WS_DISABLED = 0x8000000,
-
-            /// <summary>The window has a border of a style typically used with dialog boxes. A window with this style cannot have a title bar.</summary>
-            WS_DLGFRAME = 0x400000,
-
-            /// <summary>
-            /// The window is the first control of a group of controls. The group consists of this first control and all controls defined after it, up to the next control with the WS_GROUP style.
-            /// The first control in each group usually has the WS_TABSTOP style so that the user can move from group to group. The user can subsequently change the keyboard focus from one control in the group to the next control in the group by using the direction keys.
-            /// You can turn this style on and off to change dialog box navigation. To change this style after a window has been created, use the SetWindowLong function.
-            /// </summary>
-            WS_GROUP = 0x20000,
-
-            /// <summary>The window has a horizontal scroll bar.</summary>
-            WS_HSCROLL = 0x100000,
-
-            /// <summary>The window is initially maximized.</summary>
-            WS_MAXIMIZE = 0x1000000,
-
-            /// <summary>The window has a maximize button. Cannot be combined with the WS_EX_CONTEXTHELP style. The WS_SYSMENU style must also be specified.</summary>
-            WS_MAXIMIZEBOX = 0x10000,
-
-            /// <summary>The window is initially minimized.</summary>
-            WS_MINIMIZE = 0x20000000,
-
-            /// <summary>The window has a minimize button. Cannot be combined with the WS_EX_CONTEXTHELP style. The WS_SYSMENU style must also be specified.</summary>
-            WS_MINIMIZEBOX = 0x20000,
-
-            /// <summary>The window is an overlapped window. An overlapped window has a title bar and a border.</summary>
-            WS_OVERLAPPED = 0x0,
-
-            ///// <summary>The window is an overlapped window.</summary>
-            //WS_OVERLAPPEDWINDOW = WS_OVERLAPPED | WS_CAPTION | WS_SYSMENU | WS_SIZEFRAME | WS_MINIMIZEBOX | WS_MAXIMIZEBOX,
-
-            /// <summary>The window is a pop-up window. This style cannot be used with the WS_CHILD style.</summary>
-            WS_POPUP = 0x80000000u,
-
-            ///// <summary>The window is a pop-up window. The WS_CAPTION and WS_POPUPWINDOW styles must be combined to make the window menu visible.</summary>
-            //WS_POPUPWINDOW = WS_POPUP | WS_BORDER | WS_SYSMENU,
-
-            /// <summary>The window has a sizing border.</summary>
-            WS_SIZEFRAME = 0x40000,
-
-            /// <summary>The window has a window menu on its title bar. The WS_CAPTION style must also be specified.</summary>
-            WS_SYSMENU = 0x80000,
-
-            /// <summary>
-            /// The window is a control that can receive the keyboard focus when the user presses the TAB key.
-            /// Pressing the TAB key changes the keyboard focus to the next control with the WS_TABSTOP style.
-            /// You can turn this style on and off to change dialog box navigation. To change this style after a window has been created, use the SetWindowLong function.
-            /// For user-created windows and modeless dialogs to work with tab stops, alter the message loop to call the IsDialogMessage function.
-            /// </summary>
-            WS_TABSTOP = 0x10000,
-
-            /// <summary>The window is initially visible. This style can be turned on and off by using the ShowWindow or SetWindowPos function.</summary>
-            WS_VISIBLE = 0x10000000,
-
-            /// <summary>The window has a vertical scroll bar.</summary>
-            WS_VSCROLL = 0x200000
-        }
-
-        [Flags]
-        public enum WindowStylesEx : uint
-        {
-            /// <summary>
-            /// Specifies that a window created with this style accepts drag-drop files.
-            /// </summary>
-            WS_EX_ACCEPTFILES = 0x00000010,
-
-            /// <summary>
-            /// Forces a top-level window onto the taskbar when the window is visible.
-            /// </summary>
-            WS_EX_APPWINDOW = 0x00040000,
-
-            /// <summary>
-            /// Specifies that a window has a border with a sunken edge.
-            /// </summary>
-            WS_EX_CLIENTEDGE = 0x00000200,
-
-            /// <summary>
-            /// Windows XP: Paints all descendants of a window in bottom-to-top painting order using double-buffering. For more information, see Remarks. This cannot be used if the window has a class style of either CS_OWNDC or CS_CLASSDC.
-            /// </summary>
-            WS_EX_COMPOSITED = 0x02000000,
-
-            /// <summary>
-            /// Includes a question mark in the title bar of the window. When the user clicks the question mark, the cursor changes to a question mark with a pointer. If the user then clicks a child window, the child receives a WM_HELP message. The child window should pass the message to the parent window procedure, which should call the WinHelp function using the HELP_WM_HELP command. The Help application displays a pop-up window that typically contains help for the child window.
-            /// WS_EX_CONTEXTHELP cannot be used with the WS_MAXIMIZEBOX or WS_MINIMIZEBOX styles.
-            /// </summary>
-            WS_EX_CONTEXTHELP = 0x00000400,
-
-            /// <summary>
-            /// The window itself contains child windows that should take part in dialog box navigation. If this style is specified, the dialog manager recurses into children of this window when performing navigation operations such as handling the TAB key, an arrow key, or a keyboard mnemonic.
-            /// </summary>
-            WS_EX_CONTROLPARENT = 0x00010000,
-
-            /// <summary>
-            /// Creates a window that has a double border; the window can, optionally, be created with a title bar by specifying the WS_CAPTION style in the dwStyle parameter.
-            /// </summary>
-            WS_EX_DLGMODALFRAME = 0x00000001,
-
-            /// <summary>
-            /// Windows 2000/XP: Creates a layered window. Note that this cannot be used for child windows. Also, this cannot be used if the window has a class style of either CS_OWNDC or CS_CLASSDC.
-            /// </summary>
-            WS_EX_LAYERED = 0x00080000,
-
-            /// <summary>
-            /// Arabic and Hebrew versions of Windows 98/Me, Windows 2000/XP: Creates a window whose horizontal origin is on the right edge. Increasing horizontal values advance to the left.
-            /// </summary>
-            WS_EX_LAYOUTRTL = 0x00400000,
-
-            /// <summary>
-            /// Creates a window that has generic left-aligned properties. This is the default.
-            /// </summary>
-            WS_EX_LEFT = 0x00000000,
-
-            /// <summary>
-            /// If the shell language is Hebrew, Arabic, or another language that supports reading order alignment, the vertical scroll bar (if present) is to the left of the client area. For other languages, the style is ignored.
-            /// </summary>
-            WS_EX_LEFTSCROLLBAR = 0x00004000,
-
-            /// <summary>
-            /// The window text is displayed using left-to-right reading-order properties. This is the default.
-            /// </summary>
-            WS_EX_LTRREADING = 0x00000000,
-
-            /// <summary>
-            /// Creates a multiple-document interface (MDI) child window.
-            /// </summary>
-            WS_EX_MDICHILD = 0x00000040,
-
-            /// <summary>
-            /// Windows 2000/XP: A top-level window created with this style does not become the foreground window when the user clicks it. The system does not bring this window to the foreground when the user minimizes or closes the foreground window.
-            /// To activate the window, use the SetActiveWindow or SetForegroundWindow function.
-            /// The window does not appear on the taskbar by default. To force the window to appear on the taskbar, use the WS_EX_APPWINDOW style.
-            /// </summary>
-            WS_EX_NOACTIVATE = 0x08000000,
-
-            /// <summary>
-            /// Windows 2000/XP: A window created with this style does not pass its window layout to its child windows.
-            /// </summary>
-            WS_EX_NOINHERITLAYOUT = 0x00100000,
-
-            /// <summary>
-            /// Specifies that a child window created with this style does not send the WM_PARENTNOTIFY message to its parent window when it is created or destroyed.
-            /// </summary>
-            WS_EX_NOPARENTNOTIFY = 0x00000004,
-
-            /// <summary>
-            /// Combines the WS_EX_CLIENTEDGE and WS_EX_WINDOWEDGE styles.
-            /// </summary>
-            WS_EX_OVERLAPPEDWINDOW = WS_EX_WINDOWEDGE | WS_EX_CLIENTEDGE,
-
-            /// <summary>
-            /// Combines the WS_EX_WINDOWEDGE, WS_EX_TOOLWINDOW, and WS_EX_TOPMOST styles.
-            /// </summary>
-            WS_EX_PALETTEWINDOW = WS_EX_WINDOWEDGE | WS_EX_TOOLWINDOW | WS_EX_TOPMOST,
-
-            /// <summary>
-            /// The window has generic "right-aligned" properties. This depends on the window class. This style has an effect only if the shell language is Hebrew, Arabic, or another language that supports reading-order alignment; otherwise, the style is ignored.
-            /// Using the WS_EX_RIGHT style for static or edit controls has the same effect as using the SS_RIGHT or ES_RIGHT style, respectively. Using this style with button controls has the same effect as using BS_RIGHT and BS_RIGHTBUTTON styles.
-            /// </summary>
-            WS_EX_RIGHT = 0x00001000,
-
-            /// <summary>
-            /// Vertical scroll bar (if present) is to the right of the client area. This is the default.
-            /// </summary>
-            WS_EX_RIGHTSCROLLBAR = 0x00000000,
-
-            /// <summary>
-            /// If the shell language is Hebrew, Arabic, or another language that supports reading-order alignment, the window text is displayed using right-to-left reading-order properties. For other languages, the style is ignored.
-            /// </summary>
-            WS_EX_RTLREADING = 0x00002000,
-
-            /// <summary>
-            /// Creates a window with a three-dimensional border style intended to be used for items that do not accept user input.
-            /// </summary>
-            WS_EX_STATICEDGE = 0x00020000,
-
-            /// <summary>
-            /// Creates a tool window; that is, a window intended to be used as a floating toolbar. A tool window has a title bar that is shorter than a normal title bar, and the window title is drawn using a smaller font. A tool window does not appear in the taskbar or in the dialog that appears when the user presses ALT+TAB. If a tool window has a system menu, its icon is not displayed on the title bar. However, you can display the system menu by right-clicking or by typing ALT+SPACE.
-            /// </summary>
-            WS_EX_TOOLWINDOW = 0x00000080,
-
-            /// <summary>
-            /// Specifies that a window created with this style should be placed above all non-topmost windows and should stay above them, even when the window is deactivated. To add or remove this style, use the SetWindowPos function.
-            /// </summary>
-            WS_EX_TOPMOST = 0x00000008,
-
-            /// <summary>
-            /// Specifies that a window created with this style should not be painted until siblings beneath the window (that were created by the same thread) have been painted. The window appears transparent because the bits of underlying sibling windows have already been painted.
-            /// To achieve transparency without these restrictions, use the SetWindowRgn function.
-            /// </summary>
-            WS_EX_TRANSPARENT = 0x00000020,
-
-            /// <summary>
-            /// Specifies that a window has a border with a raised edge.
-            /// </summary>
-            WS_EX_WINDOWEDGE = 0x00000100
-        }
-
-        [Flags]
-        public enum SendMessageTimeoutFlags : uint
-        {
-            SMTO_NORMAL = 0x0,
-            SMTO_BLOCK = 0x1,
-            SMTO_ABORTIFHUNG = 0x2,
-            SMTO_NOTIMEOUTIFNOTHUNG = 0x8,
-            SMTO_ERRORONEXIT = 0x20
-        }
-
-        [DllImport("gdi32.dll", EntryPoint = "BitBlt")]
-        public static extern bool BitBlt(IntPtr hdcDest, int xDest, int yDest, int wDest, int hDest, IntPtr hdcSource, int xSrc, int ySrc, int RasterOp);
-
-        [DllImport("gdi32.dll", EntryPoint = "CreateCompatibleBitmap")]
-        public static extern IntPtr CreateCompatibleBitmap(IntPtr hdc, int nWidth, int nHeight);
-
-        [DllImport("gdi32.dll", EntryPoint = "CreateCompatibleDC")]
-        public static extern IntPtr CreateCompatibleDC(IntPtr hdc);
-
-        [DllImport("gdi32.dll", EntryPoint = "DeleteDC")]
-        public static extern IntPtr DeleteDC(IntPtr hDc);
-
-        [DllImport("gdi32.dll", EntryPoint = "DeleteObject")]
-        public static extern IntPtr DeleteObject(IntPtr hDc);
-
-        [DllImport("dwmapi.dll")]
-        public static extern int DwmEnableBlurBehindWindow(IntPtr hWnd, ref BbStruct blurBehind);
-
-        [DllImport("DwmApi.dll")]
-        public static extern int DwmExtendFrameIntoClientArea(IntPtr hwnd, ref Margins pMarInset);
-
-        [DllImport("dwmapi.dll", EntryPoint = "#127", PreserveSig = false)]
-        public static extern void DwmGetColorizationParameters(out DWM_COLORIZATION_PARAMS parameters);
-
-        [DllImport("dwmapi.dll", PreserveSig = false)]
-        public static extern bool DwmIsCompositionEnabled();
-
-        [DllImport("dwmapi.dll")]
-        public static extern int DwmIsCompositionEnabled(out bool enabled);
-
-        [DllImport("dwmapi.dll", EntryPoint = "#131", PreserveSig = false)]
-        public static extern void DwmSetColorizationParameters(ref DWM_COLORIZATION_PARAMS parameters, long uUnknown);
-
-        [DllImport("dwmapi.dll")]
-        public static extern int DwmSetWindowAttribute(IntPtr hwnd, int attr, ref int attrValue, int attrSize);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool EnumWindows(EnumWindowsProc lpEnumFunc, IntPtr lParam);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool EnumChildWindows(IntPtr parentHandle, EnumWindowsProc lpEnumFunc, IntPtr lParam);
-
-        [DllImport("user32.dll", SetLastError = true)]
-        public static extern IntPtr FindWindow(string lpClassName, string lpWindowName);
-
-        [DllImport("user32.dll", SetLastError = true)]
-        public static extern IntPtr FindWindowEx(IntPtr parentHandle, IntPtr childAfter, string className, IntPtr windowTitle);
-
-        [DllImport("user32.dll")]
-        public static extern uint GetClassLong(IntPtr hWnd, int nIndex);
-
-        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        public static extern int GetClassName(IntPtr hWnd, StringBuilder lpClassName, int nMaxCount);
-
-
-        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        public static extern int GetWindowText(IntPtr hWnd, StringBuilder lpWindowText, int nMaxCount);
-
-        [DllImport("user32.dll", EntryPoint = "GetDC")]
-        public static extern IntPtr GetDC(IntPtr ptr);
-
-        [DllImport("user32.dll")]
-        public static extern IntPtr GetDCEx(IntPtr hWnd, IntPtr hrgnClip, DeviceContextValues flags);
-
-        [DllImport("user32.dll", EntryPoint = "GetDesktopWindow")]
-        public static extern IntPtr GetDesktopWindow();
-
-        [DllImport("user32.dll")]
-        public static extern IntPtr GetForegroundWindow();
-
-        [DllImport("user32.dll", ExactSpelling = true, CharSet = CharSet.Auto)]
-        public static extern IntPtr GetParent(IntPtr hWnd);
-
-        [DllImport("user32.dll")]
-        public static extern IntPtr GetShellWindow();
-
-        [DllImport("user32.dll", EntryPoint = "GetSystemMetrics")]
-        public static extern int GetSystemMetrics(int abc);
-
-        [DllImport("user32.dll", EntryPoint = "GetWindowDC")]
-        public static extern IntPtr GetWindowDC(Int32 ptr);
-
-        [DllImport("user32.dll")]
-        public static extern bool IsWindowVisible(IntPtr hWnd);
-
-        [DllImport("user32.dll")]
-        public static extern bool RedrawWindow(IntPtr hWnd, [In] ref RECT lprcUpdate, IntPtr hrgnUpdate, RedrawWindowFlags flags);
-
-        [DllImport("user32.dll")]
-        public static extern bool RedrawWindow(IntPtr hWnd, IntPtr lprcUpdate, IntPtr hrgnUpdate, RedrawWindowFlags flags);
-
-        [DllImport("user32.dll", EntryPoint = "ReleaseDC")]
-        public static extern IntPtr ReleaseDC(IntPtr hWnd, IntPtr hDc);
-
-        [DllImport("gdi32.dll", EntryPoint = "SelectObject")]
-        public static extern IntPtr SelectObject(IntPtr hdc, IntPtr bmp);
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto)]
-        public static extern IntPtr SendMessage(IntPtr hWnd, UInt32 Msg, int wParam, IntPtr lParam);
-
-        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        public static extern IntPtr SendMessageTimeout(IntPtr windowHandle, uint Msg, IntPtr wParam, IntPtr lParam, SendMessageTimeoutFlags flags, uint timeout, out IntPtr result);
-
-        [DllImport("user32.dll", SetLastError = true)]
-        public static extern IntPtr SetParent(IntPtr hWndChild, IntPtr hWndNewParent);
-
-        [DllImport("kernel32.dll")]
-        public static extern bool SetProcessWorkingSetSize(IntPtr handle, int minimumWorkingSetSize, int maximumWorkingSetSize);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool EnableWindow(IntPtr hWnd, bool bEnable);
-       
-
-        /// <summary>
-        /// Changes an attribute of the specified window. The function also sets the 32-bit (long) value at the specified offset into the extra window memory.
-        /// </summary>
-        /// <param name="hWnd">A handle to the window and, indirectly, the class to which the window belongs..</param>
-        /// <param name="nIndex">GWL_EXSTYLE, GWL_HINSTANCE, GWL_ID, GWL_STYLE, GWL_USERDATA, GWL_WNDPROC </param>
-        /// <param name="dwNewLong">The replacement value.</param>
-        /// <returns>If the function succeeds, the return value is the previous value of the specified 32-bit integer.
-        /// If the function fails, the return value is zero. To get extended error information, call GetLastError. </returns>
-        [DllImport("user32.dll")]
-        public static extern int SetWindowLong(IntPtr hWnd, WindowLongFlags nIndex, int dwNewLong);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool SetWindowPos(IntPtr hWnd, IntPtr hWndInsertAfter, int X, int Y, int cx, int cy, SetWindowPosFlags uFlags);
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto)]
-        public static extern Int32 SystemParametersInfo(UInt32 action, UInt32 uParam, string vParam, UInt32 winIni);
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct BbStruct //Blur Behind Structure
-        {
-            public BbFlags Flags;
-            public bool Enable;
-            public IntPtr Region;
-            public bool TransitionOnMaximized;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct DWM_COLORIZATION_PARAMS
-        {
-            public int clrColor;
-            public int clrAfterGlow;
-            public int nIntensity;
-            public int clrAfterGlowBalance;
-            public int clrBlurBalance;
-            public int clrGlassReflectionIntensity;
-            public bool fOpaque;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct Margins
-        {
-            public int cxLeftWidth;      // width of left border that retains its size
-            public int cxRightWidth;     // width of right border that retains its size
-            public int cyTopHeight;      // height of top border that retains its size
-            public int cyBottomHeight;   // height of bottom border that retains its size
-        };
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct RECT
-        {
-            public int Left, Top, Right, Bottom;
-
-            public RECT(int left, int top, int right, int bottom)
-            {
-                Left = left;
-                Top = top;
-                Right = right;
-                Bottom = bottom;
-            }
-
-            public RECT(System.Drawing.Rectangle r)
-                : this(r.Left, r.Top, r.Right, r.Bottom)
-            {
-            }
-
-            public int X
-            {
-                get { return Left; }
-                set { Right -= (Left - value); Left = value; }
-            }
-
-            public int Y
-            {
-                get { return Top; }
-                set { Bottom -= (Top - value); Top = value; }
-            }
-
-            public int Height
-            {
-                get { return Bottom - Top; }
-                set { Bottom = value + Top; }
-            }
-
-            public int Width
-            {
-                get { return Right - Left; }
-                set { Right = value + Left; }
-            }
-
-            public System.Drawing.Point Location
-            {
-                get { return new System.Drawing.Point(Left, Top); }
-                set { X = value.X; Y = value.Y; }
-            }
-
-            public System.Drawing.Size Size
-            {
-                get { return new System.Drawing.Size(Width, Height); }
-                set { Width = value.Width; Height = value.Height; }
-            }
-
-            public static implicit operator System.Drawing.Rectangle(RECT r)
-            {
-                return new System.Drawing.Rectangle(r.Left, r.Top, r.Width, r.Height);
-            }
-
-            public static implicit operator RECT(System.Drawing.Rectangle r)
-            {
-                return new RECT(r);
-            }
-
-            public static bool operator ==(RECT r1, RECT r2)
-            {
-                return r1.Equals(r2);
-            }
-
-            public static bool operator !=(RECT r1, RECT r2)
-            {
-                return !r1.Equals(r2);
-            }
-
-            public bool Equals(RECT r)
-            {
-                return r.Left == Left && r.Top == Top && r.Right == Right && r.Bottom == Bottom;
-            }
-
-            public override bool Equals(object obj)
-            {
-                if (obj is RECT)
-                    return Equals((RECT)obj);
-                else if (obj is System.Drawing.Rectangle)
-                    return Equals(new RECT((System.Drawing.Rectangle)obj));
-                return false;
-            }
-
-            public override int GetHashCode()
-            {
-                return ((System.Drawing.Rectangle)this).GetHashCode();
-            }
-
-            public override string ToString()
-            {
-                return string.Format(System.Globalization.CultureInfo.CurrentCulture, "{{Left={0},Top={1},Right={2},Bottom={3}}}", Left, Top, Right, Bottom);
-            }
-        }
-    }
-
-}
