+++ Windows/DesktopWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="SchoolWallpaperChanger.External_Functions.Window1"
+﻿<Window x:Class="SchoolWallpaperChanger.Windows.DesktopWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SchoolWallpaperChanger.External_Functions"
+        xmlns:local="clr-namespace:SchoolWallpaperChanger.Functions"
         mc:Ignorable="d"
-        Title="Window1" Height="450" Width="800">
+        Title="DesktopWindow" Height="450" Width="800">
     <Grid>
         <Image x:Name="img" Stretch="Fill" Visibility="Visible" Initialized="img_Initialized"/>
     </Grid>
